Added whitelines, also made changes according to comments
