+++ deploy/etc/deployments/zgc_cluster_config.yaml
@@ -1,3 +1,10 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2020 The Authors.
+
+# Authors: Rio Zhu <@zzxgzgz>
+#
+# Summary: ConfigMap to store zgc cluster's settings.
+
 apiVersion: v1
 kind: ConfigMap
 metadata:
@@ -6,3 +13,4 @@ metadata:
     name: zgc-cluster-config
 data:
   cidr: 10.0.0.0/24
+  MAC_PRIVATE: "0x828300000000"
