+++ test/yaml/droplet.yaml
@@ -16,3 +16,4 @@ spec:
   - "10.0.0.2"
   interface: "eth0"
   status: "Init"
+
